refactor(ViewUtils): 기본 입,출력 메소드를 분리하기 위해 ViewUtils 생성
